feat: set timeout for handlers if Request-Timeout header is set.
Usually the handler context is canceled when the connection is closed or
the requests are canceled (using h2). As due to infrastructure in
between it may happen that the connection is broken, but not closed.
Therefore if the caller indicates a timeout for the operation using the
Request-Timeout header, we will apply this timeout for the handler.